New methods for interacting with database
A bit general. The additional methods included in this commit are meant to
make it easier to to deal with "missing primary keys" with many of the
methods. For example, to check whether or not a record exists, as defined
by the primary keys

Also redefining the schema for tables. Originally, due to SQLite's limited
datatype support compared to Postgres, floats were stored as integers. By
changing to Postgres, I changed back to decimals, and included additional
columns to capture datetime changes